Age Group Distribution of Influenza Positive Tests Reported by Public Health Laboratories

Retrieves the age group distribution of influenza positive tests that are reported by public health laboratories by influenza virus type and subtype/lineage. Laboratory data from multiple seasons and different age groups is provided.

age_group_distribution(years = NULL)

Arguments

years
a vector of years to retrieve data for (i.e. 2014 for CDC flu season 2014-2015). CDC has data for this API going back to 1997. Default value (NULL) means retrieve all years. NOTE: if you happen to specify a 2-digit season value (i.e. 57 == 2017-2018) the function is smart enough to retrieve by season ID vs convert that to a year.

References

CDC FluView Portal

AGD IPT Portal

Examples

agd <- age_group_distribution(years=2015)