Retrieve weekly state-level ILI indicators per-state for a given season

Retrieve weekly state-level ILI indicators per-state for a given season

ili_weekly_activity_indicators(years = NULL)

Arguments

years
a vector of years to retrieve data for (i.e. 2014 for CDC flu season 2014-2015). CDC has data for this API going back to 2008 and up until the current, active flu season. Default value (NULL) means retrieve all years. NOTE: if you happen to specify a 2-digit season value (i.e. 57 == 2017-2018) the function is smart enough to retrieve by season ID vs convert that to a year.

Note

These statistics use the proportion of outpatient visits to healthcare providers for influenza-like illness to measure the ILI activity level within a state. They do not, however, measure the extent of geographic spread of flu within a state. Therefore, outbreaks occurring in a single city could cause the state to display high activity levels.

Data collected in ILINet may disproportionately represent certain populations within a state, and therefore may not accurately depict the full picture of influenza activity for the whole state.

All summary statistics are based on either data collected in ILINet, or reports from state and territorial epidemiologists. Differences in the summary data presented by CDC and state health departments likely represent differing levels of data completeness with data presented by the state likely being the more complete.

References

ILI Activity Indicator Map Portal

Examples

# NOT RUN {
iwai <- ili_weekly_activity_indicators(2016)
# }